Refactor editor extension and post processor files under sub directories
