+++ test-vault/Example Test Note.md
@@ -6,7 +6,7 @@ The following are different clozure syntaxes. This list was copied from https://
 
 ## Paragraph Style Tests
 
-	((test]]
+((test]]
 
 Test a single {cloze} note.
